Move testDeactivatingBacklogConsumer to separate test class
